+++ src/cli/commands/mc-get-diagnostic.ts
@@ -5,6 +5,7 @@ import * as fs from "fs";
 import * as path from "path";
 import { DiagnosticInformation, MindConnectSetup } from "../..";
 import { decrypt, errorLog, homeDirLog, loadAuth, proxyLog, verboseLog } from "../../api/utils";
+import { serviceCredentialLog } from "./command-utils";
 
 export default (program: CommanderStatic) => {
     program
@@ -23,8 +24,8 @@ export default (program: CommanderStatic) => {
                     if (!options.passkey) {
                         errorLog("you have to provide a passkey (run mc gd --help for full description)", true);
                     }
-                    homeDirLog(options.verbose);
-                    proxyLog(options.verbose);
+                    homeDirLog(options.verbose, chalk.magentaBright);
+                    proxyLog(options.verbose, chalk.magentaBright);
 
                     const auth = loadAuth();
                     const setup = new MindConnectSetup(auth.gateway, decrypt(auth, options.passkey), auth.tenant);
@@ -90,17 +91,6 @@ export default (program: CommanderStatic) => {
             log(`    mc gd -k mypasskey`);
             log(`    mc get-diagnostic --config someagent.json --passkey mypasskey`);
             log(`    mc get-diagnostic --passkey mypasskey --text --all > log.csv`);
-            log(`\n  Important: \n`);
-            log(
-                `    you need to supply the ${chalk.magentaBright(
-                    "service credentials"
-                )} for this operation and provide the passkey \n`
-            );
-            log(`    how to get service credentials: `);
-            log(
-                chalk.magentaBright(
-                    `    https://developer.mindsphere.io/howto/howto-selfhosted-api-access.html#creating-service-credentials`
-                )
-            );
+            serviceCredentialLog();
         });
 };