+++ test/mindconnect-agent-3.spec.ts
@@ -200,6 +200,19 @@ describe("MindConnectApi Version 3 Agent (SHARED_SECRET)", () => {
     );
 
     it(
+        "should be able to get the currently valid agent token",
+        mochaAsync(async () => {
+            const agent = new MindConnectAgent(sharedSecretConfig);
+            if (!agent.IsOnBoarded()) {
+                await agent.OnBoard();
+            }
+
+            const currentToken = await agent.GetAgentToken();
+            currentToken.should.not.be.undefined;
+        })
+    );
+
+    it(
         "should be able to put the datasource configuration.",
         mochaAsync(async () => {
             const agent = new MindConnectAgent(sharedSecretConfig);